разбил верстку на модули и импортировал в main
