+++ src/view/all-film-list-title.js
@@ -1,0 +1,7 @@
+const createAllFilmListTitleTemplate = () => {
+  return (
+    `<h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>`
+  );
+};
+
+export {createAllFilmListTitleTemplate};
