+++ src/view/extra-films-list.js
@@ -1,0 +1,9 @@
+const createExtraFilmsListTemplate = () => {
+  return (
+    `<section class="films-list--extra">              
+              <div class="films-list__container"></div>            
+          </section>`
+  );
+};
+
+export {createExtraFilmsListTemplate};
