+++ src/view/films-board.js
@@ -1,0 +1,8 @@
+const createFilmsBoardTemplate = () => {
+  return (
+    `<section class="films">
+          </section>`
+  );
+};
+
+export {createFilmsBoardTemplate};
