+++ src/view/films-list.js
@@ -1,0 +1,9 @@
+const createFilmsListTemplate = () => {
+  return (
+    `<section class="films-list">              
+              <div class="films-list__container"></div>            
+          </section>`
+  );
+};
+
+export {createFilmsListTemplate};
