+++ src/view/load-more-btn.js
@@ -1,0 +1,7 @@
+const createLoadMoreBtnTemplate = () => {
+  return (
+    `<button class="films-list__show-more">Show more</button>`
+  );
+};
+
+export {createLoadMoreBtnTemplate};
