+++ src/view/loading-films-title.js
@@ -1,0 +1,7 @@
+const createLoadingFilmsTitleTemplate = () => {
+  return (
+    `<h2 class="films-list__title">Loading...</h2>`
+  );
+};
+
+export {createLoadingFilmsTitleTemplate};
