+++ src/view/most-commented-film-list-title.js
@@ -1,0 +1,7 @@
+const createMostCommentedFilmListTitleTemplate = () => {
+  return (
+    `<h2 class="films-list__title">Most commented</h2>`
+  );
+};
+
+export {createMostCommentedFilmListTitleTemplate};
