+++ src/view/no-data-title.js
@@ -1,0 +1,8 @@
+const createNoDataTitleTemplate = () => {
+  return (
+    `<h2 class="films-list__title">There are no movies in our database</h2>`
+  );
+};
+
+
+export {createNoDataTitleTemplate};
