+++ src/view/statistic-chart.js
@@ -1,0 +1,9 @@
+const createStatisticChart = () => {
+  return (
+    `<div class="statistic__chart-wrap">
+              <canvas class="statistic__chart" width="1000"></canvas>
+          </div>`
+  );
+};
+
+export {createStatisticChart};
