+++ src/view/statistic.js
@@ -1,0 +1,8 @@
+const createStatisticTemplate = () => {
+  return (
+    `<section class="statistic">
+          </section>`
+  );
+};
+
+export {createStatisticTemplate};
