+++ src/view/top-rated-film-list-title.js
@@ -1,0 +1,7 @@
+const createTopRatedFilmListTitleTemplate = () => {
+  return (
+    `<h2 class="films-list__title">Top rated</h2>`
+  );
+};
+
+export {createTopRatedFilmListTitleTemplate};
