+++ src/view/user-rank.js
@@ -1,0 +1,11 @@
+const createUserRankTemplate = () => {
+  return (
+    `<p class="statistic__rank">
+              Your rank
+              <img class="statistic__img" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
+              <span class="statistic__rank-label">Sci-Fighter</span>
+          </p>`
+  );
+};
+
+export {createUserRankTemplate};
